pyproject.toml
[project]
name = "aiden_companion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fellow-aiden>=0.2.2",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "reflex>=0.8.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]
